# src/api_service/openapi.yaml
# write sth here.
openapi: 3.0.3
info:
  title: SUSTech Merch Store
  description: RESTful APIs for SUSTech Merch Store.
  version: 1.0.0   

# Tags cluster API operations into different groups for fine-grained control.
tags:
  - name: Greeting
    description: Simpel welcome API
  - name: Product 
    description: Operations for getting information of products
  - name: User
    description: User account management 
  - name: Order
    description: Operations for orders

# Paths define API endpoints and operations.
paths:
  /:    
    get:
      tags:
        - Greeting
      description: Get a hello-world greeting.
      x-openapi-router-controller: src.api_service.controllers.greeting_controller
      operationId: greet
      # no request body is needed
      responses:
        '200':
          description: A successful welcome message
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  
  /products:
    get:
      tags:
        - Product
      description: Get all products on sale
      x-openapi-router-controller: src.api_service.controllers.product_controller
      operationId: list_products
      # no request body is needed
      responses:
        '200':
          description: An array of products
          content: 
            application/json: 
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/ProductBreifInfo'
  
  /products/{productId}:
    get: 
      tags:
        - Product
      description: Find product by ID
      x-openapi-router-controller: src.api_service.controllers.product_controller
      operationId: get_product_by_id
      parameters:
        - name: productId
          in: path
          required: true
          description: The ID of the product to get
          schema: 
            type: integer
      responses:
        '200':
          description: Successful operation, returns a single product.
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/ProductInfo'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/register:
    post:
      tags:
        - User
      summary: Register a new user
      x-openapi-router-controller: src.api_service.controllers.user_controller
      operationId: register_user
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User created successfully. Returns the new user's public profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '409':
          description: Conflict, the username or email is already in use.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /users/login:
    post:
      tags:
        - User
      summary: Logs user into the system
      x-openapi-router-controller: src.api_service.controllers.user_controller
      operationId: login_user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Successful login, returns an authentication token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authentication
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                  user:
                    $ref: '#/components/schemas/UserInfo'
        '401':
          description: Invalid username or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/me:
    get:
      tags:
        - User
      summary: Get current user's profile
      x-openapi-router-controller: src.api_service.controllers.user_controller
      operationId: get_current_user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation, returns the current user's profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized, authentication token is missing or invalid.
    put:
      tags:
        - User
      summary: Update current user's profile
      x-openapi-router-controller: src.api_service.controllers.user_controller
      operationId: update_current_user
      security:
        - bearerAuth: []
      requestBody:
        description: Fields to update for the user. Only non-null fields will be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully. Returns the updated user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized, authentication token is missing or invalid.
        '409':
          description: Conflict, e.g., the new email address is already in use.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 

  /users/me/delete:
    post:
      tags:
        - User
      summary: Delete the current user
      x-openapi-router-controller: src.api_service.controllers.user_controller
      operationId: deactivate_user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized, authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders:
    get:
      tags:
        - Order
      summary: Get all orders for the user
      operationId: list_orders
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of orders to return
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: Number of orders to skip for pagination
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: A list of orders for the authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderInfo'
        '401':
          description: Unauthorized, authentication token is missing or invalid.

    post:
      tags:
        - Order
      summary: Place a new order
      operationId: place_order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: A list of products and their quantities to order.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItem'
      responses:
        '201':
          description: Order created successfully. Returns the new order details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderInfo'
        '400':
          description: Bad request, e.g., product out of stock or invalid product ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized, authentication token is missing or invalid.

  /orders/{orderId}:
    get:
      tags:
        - Order
      summary: Get details of a specific order
      operationId: get_order_by_id
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: The ID of the order to retrieve.
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation, returns order details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderInfo'
        '401':
          description: Unauthorized.
        '404':
          description: Order not found or does not belong to the user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /orders/{orderId}/cancel:
    post:
      tags:
        - Order
      summary: Cancel an existing order
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: The ID of the order to cancel.
          schema:
            type: integer
      responses:
        '200':
          description: Order cancelled successfully. Returns the updated order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderInfo'
        '401':
          description: Unauthorized.
        '404':
          description: Order not found or does not belong to the user.
    

# Components define reusable data objects to be used in API operation definitions.
components:
  schemas:
    Message:
      type: object
      properties: 
        message:
          type: string
          example: "Welcome to the SUSTech Merch Store!"

    Error:
      type: object
      properties:
        message:
          type: string
          example: "Resource not found"

    ProductBreifInfo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "SUSTech Hoodie"
        price:
          type: number
          format: float
          example: 49.99
        stock:
          type: integer
          example: 500

    ProductInfo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "SUSTech Hoodie"
        description:
          type: string
          example: "🧥 A cozy, stylish hoodie featuring the official SUSTech logo, perfect for showing school spirit."
        category: 
          type: string
          example: "Apparel"
        price:
          type: number
          format: float
          example: 49.99
        stock:
          type: integer
          example: 500
        created_at:
          type: string
          format: date-time
          example: "2025-11-08T15:30:00Z"
    
    UserRegistration:
      type: object
      required:
        - sid
        - username
        - email
        - password
      properties:
        sid:
          type: string
          example: "12211418"
        username:
          type: string
          example: "loring"
        email:
          type: string
          format: email
          example: "12211418@mail.sustech.edu.cn"
        password:
          type: string
          format: password
    
    UserLogin:
      type: object
      required:
        - sid
        - password
      properties:
        sid:
          type: string
        password:
          type: string
          format: password
    
    AuthToken:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        token_type:
          type: string
          example: "bearer"

    UserInfo:
      type: object
      properties:
        id:
          type: integer
          example: 101
        sid:
          type: string
          example: "12211418"
        username:
          type: string
          example: "loring"
        email:
          type: string
          format: email
          example: "12211418@mail.sustech.edu.cn"
        created_at:
          type: string
          format: date-time
          example: "2025-11-08T10:00:00Z"

    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
          description: "New email address. Must be unique."
          example: "new.email@example.com"
        password:
          type: string
          format: password
          description: "New password. Should meet complexity requirements."
          example: "a-very-strong-new-password"
      
    OrderItem:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
          minimum: 1
    
    OrderInfo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          example: "12211418"
        product_id:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        total_price:
          type: number
          example: 49.99
        created_at:
          type: string
          format: date-time
          example: "2025-11-09T10:20:30Z"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Enter JWT Bearer token in the format: Bearer {token}"